+++ shardtree/CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.0] - PLANNED
+## [0.7.0] - 2026-07-09
 
 ### Added
 - `shardtree::LevelShifter`